Body history implemented, working on AABB Tree
