--- Unity/Assets/Scripts/TempInclude/Volatile/History/Loggers.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 4c335f2e19aa6fe49880dbde6a782ad1
-folderAsset: yes
-timeCreated: 1453248670
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
