--- Unity/Assets/Scripts/TempInclude/Volatile/History/Loggers/BodyLogger.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 729a60e48e0d9344b97fc356ca9c39fc
-timeCreated: 1453248671
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
