--- Unity/Assets/Scripts/TempInclude/Volatile/History/Loggers/ShapeLogger.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 55dae5ddf6970c94db2f208b0c53141a
-timeCreated: 1453248671
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
