--- Unity/Assets/Scripts/VolatilePhysics/AABBTree.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 158c83a0290b40740ad718d16df69e1b
-timeCreated: 1453087678
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
